requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
Pillow==10.3.0
python-multipart==0.0.9
requests==2.31.0
pytesseract==0.3.10
mangum

